pet things all in one controller
