+++ app/controllers/petController.js
@@ -20,6 +20,35 @@ exports.trackPet = function(req, callback) {
      });
 };
 
+exports.addPet = function(requestBody, callback) {
+    const user = userController.getUser().getUsername();
+    console.log("Name: " + requestBody.petname + " Type: "+ requestBody.type +
+            " Diet: " + requestBody.diet + " Walk: " + requestBody.walk);
+    var added = "INSERT INTO own_pet_belong VALUES ('" + user + "', '" + requestBody.petname + "', '" +
+                requestBody.type + "');";
+
+    if (requestBody.diet != "None") {
+        added = added + " INSERT INTO has VALUES ('" + user + "', '" + requestBody.petname + "', 'diet', '" +
+             requestBody.diet + "');";
+    }
+
+    if (requestBody.walk != "None") {
+        added = added + " INSERT INTO has VALUES ('" + user + "', '" + requestBody.petname + "', 'walk', '" +
+             requestBody.walk + "');";
+    }
+
+    const query = added;
+    console.log("Query: " + query);
+    dbController.queryGet(query, (result) => {
+        if(result.status == 200) {
+            callback(result.body.rows, "");
+        } else {
+            console.log("Failed.");
+            console.log("Status code: " + result.status);
+            callback([], result.err.message);
+        }
+    });
+}
 
 exports.editPet = function(requestBody, requestParam, callback) {
     const user = userController.getUser().getUsername();
@@ -60,7 +89,7 @@ exports.deletePet = function(req, callback) {
     console.log("Query: " + query);
     dbController.queryGet(query, (result) => {
         if(result.status == 200) {
-            callback(result.body.rows, "");
+            callback(result.body.rows);
         } else {
             console.log("Failed.");
             console.log("Status code: " + result.status);
@@ -68,3 +97,4 @@ exports.deletePet = function(req, callback) {
         }
     });
 }
+
